+++ plugins/processors/pyprocessor/srcCxx/PyDynamicData.hpp
@@ -1,3 +1,18 @@
+/******************************************************************************/
+/* (c) 2021 Copyright, Real-Time Innovations, Inc. (RTI) All rights reserved. */
+/*                                                                            */
+/* RTI grants Licensee a license to use, modify, compile, and create          */
+/* derivative works of the software solely for use with RTI Connext DDS.      */
+/* Licensee may redistribute copies of the software provided that all such    */
+/* copies are subject to this license.                                        */
+/* The software is provided "as is", with no warranty of any type, including  */
+/* any warranty for fitness for any purpose. RTI is under no obligation to    */
+/* maintain or support the software.  RTI shall not be liable for any         */
+/* incidental or consequential damages arising out of the use or inability to */
+/* use the software.                                                          */
+/*                                                                            */
+/******************************************************************************/
+
 #ifndef RTI_ROUTING_PY_DICTPRINTFORMAT_HPP_
 #define RTI_ROUTING_PY_DICTPRINTFORMAT_HPP_
 
@@ -47,7 +62,7 @@ private:
 
     public:
         PyObject *current;
-        uint64_t index;
+        size_t index;
     };
